+++ airops_weather/shared/constants.lua
@@ -1,9 +1,15 @@
 AirOpsWeather = AirOpsWeather or {}
-AirOpsWeather.Version = '0.4.0'
+AirOpsWeather.Version = '0.5.0'
 AirOpsWeather.Events = {
     requestSync = 'airops_weather:server:requestSync',
     syncState = 'airops_weather:client:syncState',
-    weatherChanged = 'airops_weather:weatherChanged'
+    weatherChanged = 'airops_weather:weatherChanged',
+    stateChanged = 'airops_weather:stateChanged',
+    profileChanged = 'airops_weather:profileChanged',
+    forecastChanged = 'airops_weather:forecastChanged',
+    overrideStarted = 'airops_weather:overrideStarted',
+    overrideEnded = 'airops_weather:overrideEnded',
+    warningsChanged = 'airops_weather:warningsChanged'
 }
 
 function AirOpsWeather.Debug(message, ...)